:white_check_mark: test: Agregar pruebas unitarias a TransactionServiceImpl

+++ src/main/java/com/bm_nttdata/transaction_ms/entity/Transaction.java
@@ -3,6 +3,8 @@ package com.bm_nttdata.transaction_ms.entity;
 import com.bm_nttdata.transaction_ms.enums.ProductTypeEnum;
 import com.bm_nttdata.transaction_ms.enums.TransactionStatusEnum;
 import com.bm_nttdata.transaction_ms.enums.TransactionTypeEnum;
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -10,8 +12,10 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.time.LocalDateTime;
-
+/**
+ * Representa una entidad de transacciones en el sistema bancario.
+ * Esta clase maneja el almacenamiento y gestión de la información de transacciones.
+ */
 @Data
 @Builder
 @NoArgsConstructor
@@ -23,8 +27,8 @@ public class Transaction {
     private String id;
     private TransactionTypeEnum transactionType;
     private ProductTypeEnum productType;
-    private Double amount;
-    private Double commission;
+    private BigDecimal amount;
+    private BigDecimal commission;
     private String sourceAccountId;
     private String targetAccountId;
     private String creditId;
